requirements.txt
pytest>=6.0
pytest-cov>=2.0
black>=21.0
flake8>=3.9
mypy>=0.900
build>=0.7.0
twine>=3.4.0
sphinx>=4.0
sphinx-rtd-theme>=1.0
sentry_sdk
PYyaml